backend/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "instagram-ai-platform-backend"
version = "0.1.0"
description = "Multi-tenant Instagram DM AI reply backend"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "fastapi>=0.111.0,<1.0.0",
    "uvicorn[standard]>=0.30.0,<1.0.0",
    "sqlalchemy[asyncio]>=2.0.31,<3.0.0",
    "asyncpg>=0.29.0,<1.0.0",
    "psycopg2-binary>=2.9.9,<3.0.0",
    "alembic>=1.13.2,<2.0.0",
    "pydantic[email]>=2.8.2,<3.0.0",
    "pydantic-settings>=2.3.4,<3.0.0",
    "python-jose[cryptography]>=3.3.0,<4.0.0",
    "passlib[bcrypt]>=1.7.4,<2.0.0",
    "bcrypt==4.0.1",
    "celery[redis]>=5.4.0,<6.0.0",
    "redis>=5.0.7,<6.0.0",
    "httpx>=0.27.0,<1.0.0",
    "python-multipart>=0.0.9,<1.0.0",
    "structlog>=24.2.0,<25.0.0",
    "slowapi>=0.1.9,<1.0.0",
    "cryptography>=42.0.0,<44.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0,<9.0.0",
    "pytest-asyncio>=0.23.8,<1.0.0",
    "pytest-mock>=3.14.0,<4.0.0",
    "httpx>=0.27.0,<1.0.0",
    "pytest-cov>=5.0.0,<6.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
